--- a/include/momentum.h
+++ b/include/momentum.h
@@ -10,4 +10,16 @@ namespace quda {
    */
   double computeMomAction(const GaugeField &mom);
 
-}
+  /**
+     Update the momentum field from the force field
+
+     mom = mom - [force]_TA
+
+     where [A]_TA means the traceless anti-hermitian projection of A
+
+     @param mom Momentum field
+     @param force Force field
+   */
+  void updateMomentum(cudaGaugeField &mom, double coeff, cudaGaugeField &force);
+
+} // namespace quda