--- a/include/tune_quda.h
+++ b/include/tune_quda.h
@@ -245,15 +245,11 @@ namespace quda {
   /**
      This derived class is for algorithms that deploy parity across
      the y dimension of the thread block.  The x threads will
-     typically correspond to the checkboarded volume, and there is no
-     explicit shared memory usage.
+     typically correspond to the checkboarded volume.
    */
   class TunableLocalParity : public Tunable {
 
   private:
-    unsigned int sharedBytesPerThread() const { return 0; }
-    unsigned int sharedBytesPerBlock(const TuneParam &) const { return 0; }
-  
     // don't tune the grid dimension
     bool tuneGridDim() const { return false; }