--- a/include/worker.h
+++ b/include/worker.h
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace quda {
+
+  class Worker {
+
+  public:
+    Worker() { }
+    virtual ~Worker() { }
+    virtual void apply(const cudaStream_t &stream) = 0;
+    
+  };
+
+};