--- a/lib/check_params.h
+++ b/lib/check_params.h
@@ -92,6 +92,16 @@ void printQudaGaugeParam(QudaGaugeParam *param) {
 
 
 #if defined INIT_PARAM
+  P(staggered_phase_type, QUDA_MILC_STAGGERED_PHASE);
+  P(staggered_phase_applied, 0);
+  P(i_mu, 0.0);
+#else
+  P(staggered_phase_type, QUDA_INVALID_STAGGERED_PHASE);
+  P(staggered_phase_applied, INVALID_INT);
+  P(i_mu, INVALID_DOUBLE);
+#endif
+
+#if defined INIT_PARAM
   P(overlap, 0);
   P(preserve_gauge, 0);
 #else