--- a/lib/dslash_events.cuh
+++ b/lib/dslash_events.cuh
@@ -11,4 +11,8 @@ namespace dslash {
   extern cudaEvent_t scatterEnd[Nstream];
   extern cudaEvent_t dslashStart;
   extern cudaEvent_t dslashEnd;
+
+  // FIX this is a hack from hell
+  // Auxiliary work that can be done while waiting on comms to finis
+  extern Worker *aux_worker;
 }