This branch is fairly wide ranging, and came from optimization of the Cineca staggered RHMC benchmark:
* More robust SVD in HISQ
* Unitarity checks now include a check for Nan
* Improved performance of multi-shift solver
  * Delay shift updates until middle of subsequent dslash to improve strong scaling
  * Stop iterating on shifted systems if iterated residuals fall below precision limit
* Optimization of HISQ and gauge force kernels through kernel fusion
* Significant of gauge force routines
* Reduced compilation time using trait mappers 
* Added `momActionQuda` for computing the momentum contribution to the action
* Added `projectSU3Quda` for projecting the gauge field onto the SU(3) manifold (uses the projection from APE smearing)
* Fixed bugs in gauge phase application kernel and added support for imaginary chemical potential (closes #284)
* In the milc interface staggered solver, use the resident gauge field until it is invalidated by constructing a new set of fat/long links
* Experimental support for fully GPU resident momentum: download when momActionQuda is computed and invalidate (disabled by default since this is not supported in mainline MILC)
* LDG loads are now enabled by default when `gauge::FloatNOrder` is used
* Improved debugging: added L1 norm reduction kernel and check for Nan when copying gauge fields (when debugging is enabled)
* Minor bug fixes
* Removal of dead code