@mikeaclark 2015-09-25
It would be great if one of you ( @alexstrel or @mathiaswagner ) would check this against the current mainline MILC.  I think it should work as is, just suboptimally.  Though I have a lot patches that are in the pipeline to make the mainline MILC.  Before I push those patches I wanted to finish some further work that will enable the gauge field to be fully resident through RHMC (as the momentum field is now with the above experimental feature).