@mathiaswagner 2015-09-25
I'll check whether I can still fire up my existing MILC tests (need to check whether I need to transfer that to different machines first in order to run it).