@mikeaclark 2015-09-29
@nmrcardoso here are some benchmarks showing the speedups of the gauge fixing with the changes I made in commit 15c9058de885dfb5be91e0f60124ee4fecc6908d to the GaugeFixQuality kernels.

Performance reported by `gauge_alg_test --recon 12`

before
```
Time: 0.906449 s, Gflop/s =  357.2, GB/s =  138.1
Time: 0.862841 s, Gflop/s =  360.6, GB/s =  133.3
Time: 1.575909 s, Gflop/s =  258.4, GB/s =   78.7
Time: 1.181364 s, Gflop/s =  341.4, GB/s =  100.6
```
after
```
Time: 0.896090 s, Gflop/s =  361.3, GB/s =  139.7
Time: 0.839907 s, Gflop/s =  370.5, GB/s =  137.0
Time: 1.536007 s, Gflop/s =  265.1, GB/s =   80.7
Time: 1.122240 s, Gflop/s =  359.4, GB/s =  105.9
```